PRSD refactor mock one login endpoint
