--- src/main/kotlin/uk/gov/communities/prsdb/webapp/local/api/oneLoginMock/LastReceivedNonce.kt
@@ -1,7 +1,0 @@
-package uk.gov.communities.prsdb.webapp.local.api.oneLoginMock
-
-class LastReceivedNonce {
-    companion object {
-        var nonce: String? = null
-    }
-}
